requirements.txt
opencv_python==4.2.0.32
Flask==1.0.2
numpy==1.18.5
Pillow
